Refactor: Move analysis classes into internal module
